[fix] テストファイルのディレクトリ構成を変更
