--- src/bbs/specs/bbs_suite_test.go
@@ -1,13 +1,0 @@
-package main_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestBbs(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Bbs Suite")
-}
